Allow INTEL_VIDEO_XPROC_SHARING to be defined
This commit fixes the use of the board flag and then the subsequent
compilation problems.

Mapping writeIntPtr to writeInt32 seems rather dubious with x64
platforms and I imagine will need to be revisited in the next
Android release.  However, this is the choice that Google made
in frameworks/av and we'll stick to it here as well.

I left this as a helper function to make the intent clearer
when it bites us later...

Change-Id: I50902ad7ff8117ef743150d3cb0f6c6f2d9033fe